refactor location of bind(), some editing for API documentation
