+++ docs/source/conf.py
@@ -45,6 +45,7 @@ epub_show_urls = "footnote"
 # -- Options for autoapi
 autoapi_type = "python"
 autoapi_dirs = ["../../src"]
+autoapi_ignore = ["*internal*"]
 autoapi_options = ["members", "special-members", "undoc-members", "imported-members"]
 autoapi_add_toctree_entry = False
 autoapi_python_class_content = "both"
