--- src/mlte/core/__init__.py
@@ -1,3 +1,0 @@
-from .bind import bind
-
-__all__ = ["bind"]
