+++ src/mlte/measurement/__init__.py
@@ -1,3 +1,4 @@
+from .bind import bind
 from .measurement import Measurement
 
-__all__ = ["Measurement"]
+__all__ = ["Measurement", "bind"]
